requirements.txt
sentence-transformers
numpy
PyYAML
faiss-cpu
flask
flask-cors
